+++ scd/src/main/java/com/bcpv/service/TipoDiabetesManager.java
@@ -11,7 +11,7 @@ public interface TipoDiabetesManager extends GenericManager<TipoDiabetes, Long> 
 	
 	void setTipoDiabetesDao(TipoDiabetesDao tipoDao);
 	
-	TipoDiabetes getTipoDiabetes (Long id_tipo);
+	TipoDiabetes getTipoDiabetes (int id_tipo);
 
     TipoDiabetes getTipoDiabetesByName (String tipo);
